LAB_04/aes_rsa_socket/requirements.txt
pycryptodome

